demos/virtual_ai_assistant_demo/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

openvino==2025.0
optimum-intel==1.22.0
optimum==1.24.0
nncf==2.15.0

llama-index==0.12.9
llama-index-llms-openvino==0.4.0
llama-index-embeddings-openvino==0.5.1
llama-index-postprocessor-openvino-rerank==0.4.1
llama-index-vector-stores-faiss==0.3.0
langchain-text-splitters==0.3.4
faiss-cpu==1.9.0

onnx==1.17.0
onnxruntime==1.20.1
torch==2.5.1

transformers==4.48.3
pymupdf==1.24.10
pyyaml==6.0.1
selfcheckgpt==0.1.7
deepeval==2.4.9

gradio==5.12.0
